requirements.txt
selenium>=4.10.0
pytest
pytest-html
python-dotenv
